docs: fix url to repo in the docsify page
The URL included at the top right side of the docsify generated page was
wrong: it linked to a non-existant address rather than to the actual
repository.

Following the instructions in the docsify library documentation, that
issue has now been solved by adding the right value to the "repo"
property in the docsify settings.